+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Added support for RPC client/server version matching through HTTP ACCEPT header (#912).
 * Added a way to ignore invalid input notes when consuming them in a transaction (#898).
 * Added `NoteUpdate` type to the note update tracker to distinguish between different types of updates (#821).
+* Updated `TonicRpcClient` and `Store` traits to be subtraits of `Send` and `Sync` (#926).
+* Updated `TonicRpcClient` and `Store` trait functions to return futures which are `Send` (#926).
 
 ### Changes